Upgrade google-cloud-sdk from 268.0.0 -> 342.0.0
Docker builds started failing because:
"Version '268.0.0-0' for 'google-cloud-sdk' was not found".

I'm not sure why they could no longer find this version of the sdk, but
after upgrading the 342.0.0 `make build` works.  We only use a couple
gcp commands in this Docker image, and I checked the GCP current documentation
for each of these commands and it LGTM.